added sections, animations, new colors, and removed deprecated css
Added an about the project section with bounce in animations on scroll.

restyled the about me section(Not a fan of how it currently looks)

removed a lot of the deprecated media query code. the page should reshape itself simple by changing the body size in the media query.

added some button hover effects and a hover effect over the contact field.